+++ src/Events/TrainEvent.java
@@ -2,7 +2,7 @@ package Events;
 
 public class TrainEvent extends EventType{
     private final boolean isEndEvent;
-    public TrainEvent(int timeStamp, boolean isEndEvent) {
+    public TrainEvent(double timeStamp, boolean isEndEvent) {
         super(timeStamp);
         this.isEndEvent = isEndEvent;
     }
