+++ src/Events/TruckQueueEvent.java
@@ -1,7 +1,7 @@
 package Events;
 
 public class TruckQueueEvent extends EventType {
-    public TruckQueueEvent(int timeStamp) {
+    public TruckQueueEvent(double timeStamp) {
         super(timeStamp);
     }
 }
